requirements_ofa.txt
# OFA 모델 환경 라이브러리 요구사항
# 생성일: 2025-08-04
# 
# 개발 환경:
# - OS: Windows 10 (64bit)
# - Python: 3.9.21
# - GPU: NVIDIA GeForce GTX 1080 Ti
# - CUDA: 11.7
# - cuDNN: 8500

# 딥러닝 프레임워크
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2+cu117

# Transformers 및 토크나이저
transformers==4.18.0.dev0
tokenizers==0.12.1

# 이미지 처리
pillow==11.2.1

# 데이터 처리
pandas==1.5.3
numpy==1.26.3

# 유틸리티
tqdm==4.67.1
matplotlib==3.9.4
seaborn==0.13.2
scikit-learn==1.1.3

# 추가 라이브러리
jsonlines>=3.1.0